Rod animation + basic drawback and casting
Updated the fishing rod model. Added animations and an animation blueprint for the fishing rod. The Rod actor now has some basic methods for receiving mouse inputs and updating in various states. The player pawn now binds the mouse click to the Rod MouseUp() and MouseDown() methods. Implemented basic logic for drawing back and casting the fishing rod by holding down and releasing the mouse click. The animation blueprint and Rod actor use a RodState enum to store and reference the current state.