Fishing line/rope proof of concept
Used a spline and an array of spline meshes to create a curved rope from a socket on the tip of the fishing rod to a pre-defined position (at the moment, (0, 0, 0). This spline rope could lead from the end of the rod to another actor or scene component, which the player would be able to cast out into the water.