Fishing line casting prototype
Added a LineEnd actor, which can be cast out in a given direction. The actor accelerates downward until it reaches 'sea level' and then stops. The PlayerRod actor now has a LineEnd child actor, and now links the rope spline from the rod end socket to the LineEnd child actor. The LineEnd is cast out when the left click is released in the drawing state. The velocity at which the LineEnd is cast out, and hence how far it travels, is dependent on both the rod's rotation and drawing alpha blend (charge) value. The LineEnd is now reset in the idle and drawing states, and the spline mesh is not re-drawn in these states eigher.