Player HUD widget, spline logic updated, reeling + random fish type testing.
Added a player HUD widget that can show and fade out a text block specifying a type of fish that has been caught.

Moved the existing rope spline logic into a CreateSplineCurved() method, and created a new method called CreateSplineStraight() which creates a spline without a curve.

If the HookState is InWater, after a random delay from casting, the PlayerRod starts reeling. When in the reeling state, holding the mouse button now turns the crank, this is done by setting the animation BlendAlpha.

Added a method to the PlayerRod actor that picks a random fish type based on the fish types' rarity values. Prototyped the caught text popping up with a randomly caught fish name, this currently pops up when entering the reeling state as the fishing minigame is not implemented yet.

The random seed is set to a system time value every tick, to ensure it is different each time the game is played.