Caught fish mechanic + minor changes
Implemented the the fish caught state, where the fish caught by the player hangs under the fishing rod. After a short delay, the player can input to return to idle/re-cast the rod. The state is currently entered after a reeling delay as the reeling minigame hasn't been implemented yet. Added a material parameter to the LineEnd's SetModel() method, and added a hook material variable to the LineEnd actor. When entering the caught state, the LineEnd model and mateiral is set to that of the fish, when returning to idle, it is re-set to the hook model and material.

Moved the player start so they are facing the lamp before their vision has adjusted to the darkness. Reduced the import size of the fish models. The curved rope dip is now a variable that can be tweaked in blueprint, as well as the caught fish height offset.