Fixed upside down error
Fixed an error allowing the player to turn the camera upside down with sporadic/fast mouse movement; the player pawn no longer caps the camera pitch after exceeding the cap, but instead pre-emptively checks if a rotation will exceed the cap. This makes it impossible to exceed the max/min looking pitch in the first place.